+++ Sources/Extensions/SwiftStdlib/ArrayExtensions.swift
@@ -111,28 +111,6 @@ public extension Array {
 		return indicies.isEmpty ? nil : indicies
 	}
 
-	/// SwifterSwift: Check if all elements in array match a conditon.
-	///
-	///		[2, 2, 4].all(matching: {$0 % 2 == 0}) -> true
-	///		[1,2, 2, 4].all(matching: {$0 % 2 == 0}) -> false
-	///
-	/// - Parameter condition: condition to evaluate each element against.
-	/// - Returns: true when all elements in the array match the specified condition.
-	public func all(matching condition: (Element) throws -> Bool) rethrows -> Bool {
-		return try !contains { try !condition($0) }
-	}
-
-	/// SwifterSwift: Check if no elements in array match a conditon.
-	///
-	///		[2, 2, 4].none(matching: {$0 % 2 == 0}) -> false
-	///		[1, 3, 5, 7].none(matching: {$0 % 2 == 0}) -> true
-	///
-	/// - Parameter condition: condition to evaluate each element against.
-	/// - Returns: true when no elements in the array match the specified condition.
-	public func none(matching condition: (Element) throws -> Bool) rethrows -> Bool {
-		return try !contains { try condition($0) }
-	}
-
 	/// SwifterSwift: Get last element that satisfies a conditon.
 	///
 	///		[2, 2, 4, 7].last(where: {$0 % 2 == 0}) -> 4